+++ Changelog.md
@@ -62,9 +62,12 @@
 - Updated Wire Textures
 
 ## 26.1.2-1.0.7_beta (Upcoming)
+- Added Cobalt Tool Suite (Powered Tools)
 - Added Copper Pipe as a fluid transport block (supports vanilla and modded fluids)
+- Added Steel Rod (Crafting Component)
 - Added wrench side-configuration support for transport blocks (pipes and wires)
 - Converted Synthetic Resin to Resin Pellet, 4 Resin Pellets can be crafted into 1 Vanilla Resin (better for balance)
+- Updated Cobalt Textures
 - Updated NeoForge version to 26.1.2.36-beta (most current, should get us past the "breaking changes")
 - Refactored "wire" to "cable" because what we have more closely resembles a cable than a wire